# storage/docs/api-docs.json
{
    "swagger": "2.0",
    "paths": {
        "/bistros": {
            "get": {
                "tags": [
                    "Bistro"
                ],
                "summary": "Get a listing of the Bistros.",
                "description": "Get all Bistros",
                "externalDocs": {
                    "description": "Some of my work can be found on this website",
                    "url": "http://www.fnkdesigns.co.uk"
                },
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/Bistro"
                                    }
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Bistro"
                ],
                "summary": "Store a newly created Bistro in storage",
                "description": "Store Bistro",
                "externalDocs": {
                    "description": "Some of my work can be found on this website",
                    "url": "http://www.fnkdesigns.co.uk"
                },
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Bistro that should be stored",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/Bistro"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Bistro"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/bistros/{id}": {
            "get": {
                "tags": [
                    "Bistro"
                ],
                "summary": "Display the specified Bistro",
                "description": "Get Bistro",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Bistro",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Bistro"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Bistro"
                ],
                "summary": "Update the specified Bistro in storage",
                "description": "Update Bistro",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Bistro",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Bistro that should be updated",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/Bistro"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Bistro"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Bistro"
                ],
                "summary": "Remove the specified Bistro from storage",
                "description": "Delete Bistro",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Bistro",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "string"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "Bistro": {
            "required": [],
            "properties": {
                "id": {
                    "description": "id",
                    "type": "integer",
                    "format": "int32"
                },
                "name": {
                    "description": "name",
                    "type": "string"
                },
                "created_at": {
                    "description": "created_at",
                    "type": "string",
                    "format": "date-time"
                },
                "updated_at": {
                    "description": "updated_at",
                    "type": "string",
                    "format": "date-time"
                }
            },
            "externalDocs": {
                "description": "Some of my work can be found on this website",
                "url": "http://www.fnkdesigns.co.uk"
            }
        },
...
    }
}